Move rest api to pbench namespace
- Moved server/lib/pbench_server to pbench/server/api
- Moved server/lib/pbench_server/pbench_server.py to
  pbench/cli/server/api.py
- Made pbench/cli/server/api.py installable via setuptools in
  setup.cfg.
- Moved tests to pbench/test/server/unit
- Updated imports.
- Updated tox.ini
