+++ pbench/test/unit/server/conftest.py
@@ -2,9 +2,9 @@ import os
 import shutil
 import pytest
 
-from pbench_server.config import ServerConfig
-from pbench_server.pbench_server import create_app
-from pbench_server.tests.common import mock_get_config_prefix
+from pbench.server.api.config import ServerConfig
+from pbench.cli.server.api import create_app
+from pbench.test.unit.server.common import mock_get_config_prefix
 
 
 @pytest.fixture(scope="session", autouse=True)
