+++ pbench/test/unit/server/test_requests.py
@@ -3,7 +3,7 @@ import os
 import pytest
 from werkzeug.utils import secure_filename
 
-from pbench_server.tests.common import (
+from pbench.test.unit.server.common import (
     filename,
     temp_dir,
 )
