+++ setup.cfg
@@ -14,5 +14,9 @@ classifier =
 [files]
 packages = pbench
 
+[entry_points]
+console_scripts = 
+   pbench-server = pbench.cli.server.api:main
+
 [tools:pytest]
 testpaths = pbench/test
