+++ tox.ini
@@ -37,12 +37,11 @@ commands =
 
 [testenv:pytest_server]
 setenv =
-  PYTHONPATH = {envdir}{:}{toxinidir}{:}{toxinidir}/server/lib/
-  _PBENCH_SERVER_CONFIG={toxinidir}/server/lib/pbench_server/tests/fixtures/config/pbench-server.cfg
+  _PBENCH_SERVER_CONFIG={toxinidir}/pbench/test/unit/server/fixtures/config/pbench-server.cfg
 commands_pre =
   find {toxinidir} -type f -not -path '{toxinidir}/.tox/*' -path '*/__pycache__/*' -name '*.py[c|o]' -delete
 commands =
-  pytest -s ./server/lib/pbench_server/tests
+  pytest -s ./pbench/test/unit/server
 
 [testenv:pytest_agent]
 setenv =
